+++ src/node/rpc/gov_effects.h
@@ -40,13 +40,5 @@ namespace ccf
     {
       impl.trigger_snapshot(tx);
     }
-
-    void trigger_acme_refresh(
-      ccf::kv::Tx& tx,
-      const std::optional<std::vector<std::string>>& interfaces =
-        std::nullopt) override
-    {
-      impl.trigger_acme_refresh(tx, interfaces);
-    }
   };
 }